content/media/video/test/test_wav_seek7.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Wave Media test: seek test 7</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript;version=1.7">
// Test for bug identified by Chris Pearce in comment 40 on
// bug 449159.
var seekCount = 0;
var completed = false;
var interval;

function startTest() {
  if (completed)
    return false;

  var seek = function () {
    var i = 0;
    for (;;) {
      i += 0.1;
      if (i > 1)
        i = 0;
      yield i;
    }
  }();
  var v = document.getElementById('v');
  interval = setInterval(function() { v.currentTime=seek.next(); }, 10);
  return false;
}

function seekEnded() {
  if (completed)
    return false;

  seekCount++;
  ok(true, "Seek " + seekCount);
  if (seekCount == 3) {
    clearInterval(interval);
    completed = true;
    SimpleTest.finish();
  }
  return false;
}

SimpleTest.waitForExplicitFinish();
</script>
</pre>
<audio id='v'
       onloadedmetadata='return startTest();'
       onseeked='return seekEnded();'>
  <source type='audio/x-wav' src='r11025_s16_c1.wav'>
</audio>
</body>
</html>
